+++ src/components/Form/CharacterCountLimitedTextArea/index.js
@@ -1,5 +1,5 @@
 import PropTypes from 'prop-types'
-import TextArea from '.'
+import TextArea from '../TextArea'
 import { useState } from 'react'
 
 const CharacterCountLimitedTextArea = ({
