+++ src/components/Form/index.js
@@ -1,8 +1,8 @@
 import Button from './Button'
-import PrimarySubmitButton from './Button/PrimarySubmitButton'
+import PrimarySubmitButton from './PrimarySubmitButton'
 import Select from './Select'
 import TextArea from './TextArea'
-import CharacterCountLimitedTextArea from './TextArea/CharacterCountLimitedTextArea'
+import CharacterCountLimitedTextArea from './CharacterCountLimitedTextArea'
 import TextInput from './TextInput'
 import DataList from './DataList'
 import Checkbox from './Checkbox'
