+++ src/components/Layout/Header/index.js
@@ -4,7 +4,7 @@ import Link from 'next/link'
 import UserContext from '../../UserContext'
 import cx from 'classnames'
 import { headerLinksForUser } from 'src/utils/headerLinks'
-import MobileMenuButton from './MobileMenuButton'
+import MobileMenuButton from '../MobileMenuButton'
 
 const HeaderComponent = ({ serviceName, toggleMobileMenu, mobileMenuOpen }) => {
   const { user } = useContext(UserContext)
